[PRO-916] Add compatibility with ActiveSupport::Cache::RedisCacheStore
This is part of an ongoing effort with upgrading to Rails 8.0. The
underlying gem used for our cache needs hasn't actually defined to give
support to the new version of ActiveSupport gem that rails uses, since
natively Rails has a very good support for it. Given that premise, using
the default mechanism for Redis support, has a different class handling,
which is the one added to the list of handlers as part of this commit